+++ packages/provider_kit/example/lib/example_kits/0_single_state/5_view_state_consumer.dart
@@ -18,7 +18,7 @@ class ViewStateConsumerExample extends StatelessWidget {
             title: "View State Consumer",
             child: ViewStateConsumer.of<ItemsProvider, List<Item>>(
               initialStateListener: () => context.showToast('initial'),
-              shouldCallListenerOnInit: true,
+              callListenerOnInit: true,
               dataStateListener: (data) => context.showToast(data.toString()),
               dataBuilder: (data) => Text(data.toString()),
             ));